Backend: Fix controllers

+++ Easy_News/Backend/Backend/DataAccess/IRepository.cs
@@ -7,7 +7,7 @@ namespace Backend.DataAccess
 {
     public interface IRepository<DBEntity, ModelEntity>
     {
-        Task<IEnumerable<ModelEntity>> Get(long id = -1);
+        Task<IEnumerable<ModelEntity>> Get(int id = -1);
         Task<ModelEntity> Insert(ModelEntity entity);
         Task<ModelEntity> Update(ModelEntity entity);
         Task<bool> Delete(long idEntity);
